# Optimize hcl report execution time.