--- a/include/hal/common-mock-func.sh
+++ b/include/hal/common-mock-func.sh
@@ -406,6 +406,7 @@ TEST_TOUCHPAD_PATH="${TEST_TOUCHPAD_PATH:-}"
 TEST_AC_PRESENT="${TEST_AC_PRESENT:-}"
 TEST_MEI_CONF_PRESENT="${TEST_MEI_CONF_PRESENT:-true}"
 TEST_INTEL_FUSE_STATUS="${TEST_INTEL_FUSE_STATUS:-0}"
+TEST_SOUND_CARD_PRESENT="${TEST_SOUND_CARD_PRESENT:-true}"
 
 fsread_tool_common_mock(){
 # This functionn emulates read hardware specific file system resources or its
@@ -443,6 +444,11 @@ fsread_tool_test_mock(){
   if [ "$_arg_e" = "/sys/class/power_supply/AC/online" ]; then
   # Emulating AC status file presence, check check_if_ac func. for more inf.:
     [ "$TEST_AC_PRESENT" = "true" ] && return 0
+  fi
+
+  if [ "$_arg_f" = "/sys/class/sound/card0/hw*/init_pin_configs" ] || [ "$_arg_f" = "/proc/asound/card0/codec#*" ]; then
+  # Emulate sound card presence, check dasharo-hcl-report for more inf.:
+    [ "$TEST_SOUND_CARD_PRESENT" = "true" ] && return 0
   fi
 
   return 1