--- a/reports/dasharo-hcl-report
+++ b/reports/dasharo-hcl-report
@@ -141,7 +141,7 @@ do
   SND_HW_FILE=`echo $SND_HW_FILES | cut -d ' ' -f 1`
   SND_CODEC_FILE=`echo $SND_CODEC_FILES | cut -d ' ' -f 1`
 
-  if [ -f "$SND_HW_FILE" ] && [ -f "$SND_CODEC_FILE" ]; then
+  if  $FSREAD_TOOL test -f "$SND_HW_FILE" && $FSREAD_TOOL test -f "$SND_CODEC_FILE"; then
     break
   else
     sleep 5