Related issue: https://github.com/Dasharo/open-source-firmware-validation/issues/652.

Results:

Without https://github.com/Dasharo/open-source-firmware-validation/pull/658 and this PR:

```bash
λ time robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v rte_ip:127.0.0.1 -v dpp_logs_key:'x' -v dpp_download_key:'x' -v dpp_password:'x' -v boot_dts_from_ipxe_shell:True -v dts_ipxe_link:http://192.168.4.175:8080/ipxe -i msi_dpp dts/dts-e2e.robot
==============================================================================
Dts-E2E                                                                       
==============================================================================
E2E003.003 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.004 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.005 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.006 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.009 MSI PRO Z690-A DDR-4 update (Coreboot + UEFI -> Coreboo... | PASS |
------------------------------------------------------------------------------
E2E003.010 MSI PRO Z690-A update (Coreboot + UEFI -> Coreboot + UE... | PASS |
------------------------------------------------------------------------------
Dts-E2E                                                               | PASS |
6 tests, 6 passed, 0 failed
==============================================================================
Debug:   /home/danillklimuk/Projects/DTS/open-source-firmware-validation/command_log.txt
Output:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/output.xml
Log:     /home/danillklimuk/Projects/DTS/open-source-firmware-validation/log.html
Report:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/report.html
robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v  -v  -v  -v 
284.90s user 3.05s system 23% cpu 20:12.88 total
```

Without this PR:

```bash
λ time robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v rte_ip:127.0.0.1 -v dpp_logs_key:'x' -v dpp_download_key:'x' -v dpp_password:'x' -v boot_dts_from_ipxe_shell:True -v dts_ipxe_link:http://192.168.4.175:8080/ipxe -i msi_dpp dts/dts-e2e.robot
==============================================================================
Dts-E2E                                                                       
==============================================================================
E2E003.003 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.004 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.005 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.006 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.009 MSI PRO Z690-A DDR-4 update (Coreboot + UEFI -> Coreboo... | PASS |
------------------------------------------------------------------------------
E2E003.010 MSI PRO Z690-A update (Coreboot + UEFI -> Coreboot + UE... | PASS |
------------------------------------------------------------------------------
Dts-E2E                                                               | PASS |
6 tests, 6 passed, 0 failed
==============================================================================
Debug:   /home/danillklimuk/Projects/DTS/open-source-firmware-validation/command_log.txt
Output:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/output.xml
Log:     /home/danillklimuk/Projects/DTS/open-source-firmware-validation/log.html
Report:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/report.html
robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v  -v  -v  -v 
51.44s user 1.94s system 10% cpu 8:10.60 total
```

With both PRs:

```bash
λ time robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v rte_ip:127.0.0.1 -v dpp_logs_key:'x' -v dpp_download_key:'x' -v dpp_password:'x' -v boot_dts_from_ipxe_shell:True -v dts_ipxe_link:http://192.168.4.175:8080/ipxe -i msi_dpp dts/dts-e2e.robot
==============================================================================
Dts-E2E                                                                       
==============================================================================
E2E003.003 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.004 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.005 MSI PRO Z690-A DDR-4 initial deployment (legacy -> Core... | PASS |
------------------------------------------------------------------------------
E2E003.006 MSI PRO Z690-A initial deployment (legacy -> Coreboot +... | PASS |
------------------------------------------------------------------------------
E2E003.009 MSI PRO Z690-A DDR-4 update (Coreboot + UEFI -> Coreboo... | PASS |
------------------------------------------------------------------------------
E2E003.010 MSI PRO Z690-A update (Coreboot + UEFI -> Coreboot + UE... | PASS |
------------------------------------------------------------------------------
Dts-E2E                                                               | PASS |
6 tests, 6 passed, 0 failed
==============================================================================
Debug:   /home/danillklimuk/Projects/DTS/open-source-firmware-validation/command_log.txt
Output:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/output.xml
Log:     /home/danillklimuk/Projects/DTS/open-source-firmware-validation/log.html
Report:  /home/danillklimuk/Projects/DTS/open-source-firmware-validation/report.html
robot -b command_log.txt -v snipeit:no -L TRACE -v config:qemu -v  -v  -v  -v
 35.97s user 1.03s system 14% cpu 4:07.11 total
```